aws/requirements.txt
aws-cdk-lib>=2.151.0,<3
constructs>=10.0.0,<11
aws-cdk.aws-apigatewayv2-alpha>=2.151.0a0,<3
aws-cdk.aws-apigatewayv2-integrations-alpha>=2.151.0a0,<3
aws-cdk.aws-lambda-python-alpha>=2.151.0a0,<3
